work/final/app/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-MySQLdb==2.0.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
mysqlclient==2.2.6
python-dotenv==1.0.1
Werkzeug==3.1.3
